Revert "Use long instead of int"
This reverts commit 80972e01eb411c449d7949a2831cd7f5b8584c8b.

+++ HCore-Cache/Cache/Impl/RedisConnectionPoolImpl.cs
@@ -14,7 +14,7 @@ namespace HCore.Cache.Impl
 
         private readonly ICollection<ConnectionMultiplexer> _connectionMultiplexers = new List<ConnectionMultiplexer>();
 
-        private long _currentIndex = -1;
+        private int _currentIndex = -1;
 
         private bool _disposed;
